The Parser Works Now!
May still be a few bugs but for now we've got a way to parse from text into Proposition data type. I'd consider that a success for now.
Stilll need to implement NOTs or _`¦_'s here but we could start implementing the indux/redux rules maybe in a new file?